moved the vocaulary tool into a new schema cli command
