+++ c7n/cli.py
@@ -93,6 +93,16 @@ def setup_parser():
     validate.add_argument("--debug", action="store_true",
                           help="Dev Debug")
 
+    schema = subs.add_parser('schema')
+    schema.set_defaults(command=commands.schema)
+    schema.add_argument('--summarize', action="store_true",
+                        help="Summarize counts of available resources, \
+                              actions and filters")
+    schema.add_argument('--json', action="store_true",
+                        help="Switch output to JSON")
+    schema.add_argument("-v", "--verbose", action="store_true",
+                        help="Verbose Logging")
+
     #resources = subs.add_parser('resources')
     #resources.set_defaults(command=commands.resources)
     #_default_options(resources)
